Updated wsgi.py so that it works under the Django test server
Updated the base settings so that all the code packages don't need to be preceded by 'apps' (put
apps on the Python path since all the code lives in apps, no need to have to keep specifying it).
Also updated location of wsgi.py since it was moved to be where Openshift expects it to be.